Medium changes:
- Adjust free culture / free use header to Openness

- Adjust body section on resourcefulness:

X. Resourcefulness
We are relentlessly resourceful. As we grow and have ready access to capital, it is our obligation to token holders to fight bureaucracy and inefficiencies. This means solving problems in the most effective way possible at lower economic costs (in terms of capital, time and resources).

- Liberty: Add coercion-resistant sentence
...and economic freedoms. This includes being coercion-resistant.

- Add back agnostic sentence as it seems more related here than under Liberty, and has come up a few times in conversations as something these principles should capture.
II. Censorship resistance
We enable free flow of information. No content is under surveillance. We abide by the cryptoeconomic design principle of censorship resistance. Even stronger, Status is an agnostic platform for information.